requirements.txt
-- ingest data from kaggle 
kaggle

-- upload data from local to GCS
google-cloud-storage>=2.0.0


apache-airflow

psycopg2-binary
